feat: Add docker-compose-dev.yaml file
This commit adds a new `docker-compose-dev.yaml` file, which specifies the configuration for the development environment. It includes services for `nginx`, `data-viz`, `data-viz-frontend`, `data-viz-analyzer`, `data-viz-identity`, `postgres`, and `pgadmin`, along with their respective build/image settings, ports, volumes, and networks. This file allows for easier setup and management of the development environment.