Fix typo in docker service names and update Keycloak environment variables
- Fix the typo in the docker service names from `data-viz-idenity` to `data-viz-identity`
- Update the environment variables related to Keycloak in the `.env` file

PROXY_ADDRESS_FORWARDING is added to the `.env` file with a value of `true`

The nginx configuration file, `default.conf.template`, is modified to use the corrected service name `data-viz-identity` in the `proxy_pass` directive. Additionally, `proxy_set_header X-Forwarded-Port $server_port;` is added.

This commit fixes the service names and updates the Keycloak environment variables to ensure proper functioning of the application.